[OD]od-1050 Crash after switching from interact to pick mode: with a locked patch displayed only at sections on a random line
Summary:
The problem is caused by calculating distance to section in multi threads( class LockedPointsCalculator in vishorizondisplay ) in which the traces path is conflict in access by re-calculatation every time.
and also the traces path calculation takes lot of time. actually it is not necessary to re-calculate. only if random line is changed we do the calculation.
So I added a member variable to avoid re-calculation and solved this problem.

Test Plan: Tested in F3-demo many times following the steps described in Jira. it works fine.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10792