od6.0: Fixed non-deleted markerset in visSurvey::PickSetDisplay
Summary: the marker set created in PickSetDisplay::createOneMarker was referenced in that function, but not used by this class, and referenced again when given to the dragger. Since the refcount was one too many the object was never deleted. Found because visBase::MarkerSet has now a hiddenparam in od6.0 that triggers a pErrMsg.

Test Plan: no more pErrMsg from HiddenParam. No crash when either starting or closing od_main

Reviewers: ding

Reviewed By: ding

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D10808