od6.0: OD-930, Interpolation of picked horizon data does not work + crash
Summary: The horizon interpolator was either crashing or was giving undefined values. Earlier the adapter was used to fill up the array but storage was not defined and hence crash was there when the data was accessed. Now the interpolation is done on array and adapter is used to fill the datapointset with new values

Test Plan: All the algorithms are now running without crash and giving correct result

Reviewers: raman

Reviewed By: raman

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10755