[OD6.0]OD-1074 horizons with stepout larger than survey stepout are not displayed at section correctly
Summary:
This revision is ported from the revision D1074.
As discussed, we will skip the points which are not included in the displayed horizon.

Test Plan: Tested in F3-demo-d55 with special horizon: Surface_constrained_by_k2_tmp. it works fine. and it still works as before for normal horizons.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D10822