od6.0: OD-1043, Well-to-seismic tie: Update display button in Edit Time-Depth Model window doesn't work
Summary: The display of the track which is updated by picking up seeds in well tie to seismic was faulty. The display was not being updated. It was because the updated track was not provided and operation was done on the earlier generated track

Test Plan: The bug is tested in F3Demo2015 using Well F02-1

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10816