Enabled support for specifying coordinate system
Summary:
OD-259
The plan is to support the user to specify the coordintate system he is using, so that can be apropriatily converted to long/lat. Game plan in the first instance is to have a factory based system for coordinate systems. The survey will have one instance of the class, and there from to the conversion to latitude/longitude.

Typical sub-classes may later be UTM and similar. The requirement is that all classes can convert from their own x/y to Geographical coordinates on the WGS85 datum.

Step 2 is to let the user create a repository of often used coordinate systems. Such may be UTM, zone 32, but also more complex ones. The repository will be a list
of IOPars that can be used to reconstruct the coordinate system using the Coords::PositionSystem::create( iopar );

Step 2 is prepared in some API functions, but no UI and no implemementatin of the repository has been made.

Step 3 would be to allow x/y transforms when importing and exporting.

Test Plan: The survey-info reading and writing is tested. The UnlocatedXY is tested for reading, writing usepar and editing in UI.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10858