[OD6.0]OD-1048 Memory leak issue in using Polygon selection tool.
Summary:
The revision is ported from revision D10855.
When doing unref of osggeo polygon selection, it doesn't work, because the ref count is not correct due to addChild() in which it is ref() again.
So the solution we have to do removeChild and redo ref.

Test Plan: Tested in F3-demo with revision D10855 together. It works fine.

Reviewers: ranojay

Reviewed By: ranojay

Differential Revision: https://phabricator.dgbes.com:8443/D10857