od6.0: Fixed order of 2D lines in selection box
Summary:
When adding 2D lines to the 3D scene, the order is the same as the order of Geometry omf file. Changed this to alphabetical order.
Not sure if there's a better way of sorting this list.

Test Plan: Tested on nlog_2D survey. Order is correct and I don't see a delay in popup.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10845