od6.0: Fix for crash in setPickDirs
Summary: Fix for crash in setPickDirs as reported in OD-993. Reproducible when all 3D  scenes are closed and a 2D viewer is present.

Test Plan:
Started OD, closed 3D scene. Add a 2D viewer and a pickset, then select the pickset option 'Set Directions'. No crash.
Also tried with a 3D scene up. No crash either.
Direction of markers not good, as expected with a 0 velocity.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D10838