od6.0: Removed a call to commitToTracker
Summary:
The callstack of the crash in OD-840 shows a call to commitToTracker from the trackerWinClosedCB function. This call is not needed anymore, so removed it for safety reasons.
Could not find the exact reason for the crash itself, but this might solve it.

Test Plan: Started tracking, removed horizon to trigger a setup dialog close. All fine, no crash.

Reviewers: yuancheng

Reviewed By: yuancheng

Differential Revision: https://phabricator.dgbes.com:8443/D10860