od6.0: Added toolbutton to process attribute
Summary:
As a shortcut, added a toolbutton to process the selected attribute. Also a new setInput was needed in uiAttrVolOut to preset the selected attribute.
Implements OD-991

Test Plan: Opened a default attributeset and clicked on the new button for several attributes. Gui is filled in correctly. Clicking on OK, will start the process as expected.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D10859