od: Fix for crash in hilbert calculation of null traces
Summary:
Got a crash when calculating anything with a hilbert transform. Trace only had zeros. After removing this check, the crash was gone.
Not sure if this is the right solution. Perhaps we need to fill 'output' with zeros when the sum==0?

Test Plan: Tested in Kerry3D survey, inline 653. No crash anymore when trying to display instantaneous attributes.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D10848