OD6.0: OD-1042, Well-to-seismic tie: change default marker size to 2
Summary: The marker size was set to default size of 15 for new surveys. This has now being changed to 2.

Test Plan: Created a new survey by manually feeding the parameters of F3Demo2015 and importing seismics and well properites. The default size initially was 15 after the changes it was set to 2

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10876