[OD]OD-526,OD-906 3D Visualization of big horizon
Summary:
Bug 526 and 906 complain the 3d visualization of big horizon. We have done several tests in big horizons with 9000 by 9000 in-lines and cross lines. It is indeed it will take 
more than 15 GB memory to load horizon and construct horizon rendering. if graphical memory is small, it will take very long time to display it.

This revision tries to improve this in two aspects:

1. As suggested from Nanne, in uivis, we will show a message to client when a horizon is too big, and recommend clients to reduce horizon resolution, 
so that it can save a certain memory. This is presented in visuiemobject.cc

2. in visbase, We added and modified the existing code creating horizon tile. The changed code allow to create empty tiles and stored in Array2DImpl. This solution can save memory in 
the case of horizon having the area without data ( it is very often happened according to the experience from case study people ).

Test Plan:
The revision is tested in a fake survey (manually created survey) which almost contains 8000 by 9000 in lines and cross lines. We choose horizon with the resolution 1 by 1,
It will generates 17934 tiles. and we manually remove the area which occupies 15000 tiles, and we test how is memory in the new code.

for the old code, after loading horizon data and visualized the horizon, the memory is reduced 8.3 GB. Manipulating scene is difficult( such as rotation, zoom/in).
for the new code, after loading horizon data and visualized the horizon, the memory is reduced 2.5 GB. Manipulating scene is relative easy.

Obviously, this revision can save a certain memory only in the case of some area has no data. If a huge horizon has full data in the area, we need figure out another solution.

I did other tests in F3-demo with all normal horizon functions, it works as before.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D10878