[OD6.0]OD-526,OD-906 3D Visualization of big horizon
Summary:
This revision is ported from the revision D10878.
The solution is only for short term.

Test Plan:
The revision is tested in a fake survey (manually created survey) which almost contains 8000 by 9000 in lines and cross lines. We choose horizon with the resolution 1 by 1,
It will generates 17934 tiles. and we manually remove the area which occupies 15000 tiles, and we test how is memory in the new code.

for the old code, after loading horizon data and visualized the horizon, the memory is reduced 8.3 GB. Manipulating scene is difficult( such as rotation, zoom/in).
for the new code, after loading horizon data and visualized the horizon, the memory is reduced 2.5 GB. Manipulating scene is relative easy.

Obviously, this revision can save a certain memory only in the case of some area has no data. If a huge horizon has full data in the area, we need figure out another solution.
I did other tests in F3-demo with all normal horizon functions, it works as before.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D10884