od: Fixed Sequential reader not reading a second time the data
Summary:
The trace reader initialization was removed from SequentialReader::setDataPack implementation when fixing OD-299 (commit from May 18, 2016). Without it the confidence datapack was not read anymore in the unconformity tracker and HorizonCube processing. That is because a same reader is used twice to fill two datapacks, since the dip components goes to one datapack while the confidence goes to another.
od specific, not in od6.0
In general one should start using the setDataPack function for any usage of this reader, and remove all cases where the sequential reader creates and owns the datapack to be filled, because the next step after reading will be to transfer that owned datapack into another before the reader gets out of scope, which is a large waste of memory and time ressources.

Test Plan: ran a HorizonCube batch process. Checked in gdb the datapack values after reading.

Reviewers: raman, bert

Reviewed By: bert

Subscribers: nanne, bert

Differential Revision: https://phabricator.dgbes.com:8443/D10896