Simplified Coord and Coord3
Summary:
I want to (eventually) make a Coord3f for float precision coordinates in the
visualization. Hence I wanted to reuse the code of Coord3, but I could not.

Hence I made Coord3 a typedef of Geom::Point3<double>. I also modernized the
API by adding _ after variable names.

Further, same changes on Coord. Some functions became template functions, such as abs, distTo, as it is the caller who knows what precision that is used.

Adapted a few places where it was obvious that double precision was not needed.

Full review not requested, just that you are on board with the idea.

Test Plan: OpendTect compiles and starts.

Reviewers: nanne

Reviewed By: nanne

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D10894