od6.0: Added macAddressHash
Summary: Added new function macAddressHash to replace uniqueSystemID. The new function will filter out the virtualbox mac address if present (OD-634). As the return value is not a unique id, a function name change was needed.

Test Plan: Ran the test program to verify correct retrieval of the mac address.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D10914