od6.0: Fixed error in readonly definition of comboboxes
Summary:
The setReadOnly function in uiComboBox was setting the editable property, which is something different. The readonly state can be set on the lineedit part of the combobox.
Fixed this by changing setReadOnly and adding setEditable.

Found this after Mike Pelissier told us he could not copy the text of the attribute inputs anymore.

Test Plan: Tested in uiAttrSel. This combobox is now readonly (entries can not be changed) but they're also editable (entries can be selected and copied).

Reviewers: yuancheng

Reviewed By: yuancheng

Differential Revision: https://phabricator.dgbes.com:8443/D10933