od6.0: OD-876, Wheeler 2D attribute listed for display in Time scene
Summary: In time scene while selecting the 2D attribute other domain like wheeler were also shown. This has been fixed. An empty zdomain key is provided for search, when wheeler is returned the name is not saved, else the return value is null

Test Plan: The fix has been tested in F3Demo2015 and is working

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10697