od6.0: Added link to training manual in Help menu
Summary:
See title and as requested in OD-1146
Also changed the link name of the workflows manual to HTML_WF which seems to be the new format.

Please verify with Jan if this training manual addition is OK.

Test Plan: Clicked on both, How-to and training manual items in Help menu. Correct pages are openened in browser.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10932