od6.0: Added 'Other plugins' Toolbar
Summary:
As discussed with Marketing, we now have a ToolBar called 'Other Plugins". It will be initialized when it is invoked for the first time.

There is one issue with the Toolbar layout though. For a new user, in the default layout we have only the dtectTB on the first rwo and the Context Toolbar inserts a newline sending all the plugin Toolbars to the second row. If we delay the creation of Context Toolbar, we can accommodate the plugins (at least dGB Plugins) toolbar on the first row.

Test Plan: Tested by temporarily adding the GMT button to this new toolbar. It appears next to the dGB toolbar. Also, toggle on/off from view menu works fine.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10941