OD6.0: OD-286, In the Edit log window the remove row option does not work
Summary: The was an option to delete single\multiple rows but on executing it rows were deleted temporarily from the dialog and were stored back when dialog was reopened. Also the MD values were not allowed to be changed and if tried to change the MD values the inserted value being read as log value. Also row insertion was not working properly.

Test Plan: The row single\multiple are being deleted. Selection may not be continous. The MD values are being changed now and stored properly. Also on row insertion the values are now being properly saved.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D10942