od6.0: Fix for Gridding horizons with sparse points [OD-721]
Summary: If a horizon has Inl/Xl steps larger than those of the survey, it results in a crappy display in Basemap and 2D viewer. The reason was while creating the XY-grid from the Inl/Xl-grid all those XY points that transformed to a BinID not conforming to the horison step were skipped(filled with undef). Now they take into account the nearest available BinIDs. This was achieved by snapping the transformed BinID to the TrcKeySampling of the horizon.

Test Plan: Created a copy of a 3D hor in F3 using steps 4X4. The display in Basemap as well as the 2D viewer looks smooth after the fix.

Reviewers: satyaki

Reviewed By: satyaki

Differential Revision: https://phabricator.dgbes.com:8443/D10996