[OD]OD-288 PDF3D-3D: support faults, part1: od interfaces
Summary:
This solution is a set of interfaces to provide necessary information to be used in plug-in: dgbPDF3D.
Note: in visShape, getOsgNormal() is only used in dgbPDF3D. I changed the transformation, because I think osg normal is just the 
true normal being used in display. The name getOsgNormal, it means to get true normal used in current display. So I removed the transform issues.

Test Plan: Tested with new dgbPDF3D together, it works fine.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D10989