od: StepInterval<typ>::isCompatible does not work properly for float case
Summary: the default epsilon used in the last statment was far too small for the float case, it should be at most 1e-5 since float have a precision of 1e-6. In the implementation of the macro casteps in the last statment is a relative epsilon, thus can be hardcoded as macro argument.

Test Plan: ranges.h does not have a test program, thus abused test_trckeyzsampling to add a corresponding test for float and double. Fails without the changes, works after.

Reviewers: kristofer, bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D10997