[OD6.0]OD-194 Tacking: make Seed Index value unique (currently it comes from seed XL position)
Summary:
This revision is ported from revision D10947 in od trunk. 

The next step we are going to save seedid on disk to keep the id no changing when reload horizon. 
but it will be implemented in different Jira issue.

Test Plan: Tested in F3-demo, it works fine. all seeds id are unique number.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10964