od: Crash during insertion of row prior to first row
Summary: While inserting a row prior to first row in well log editor (well manager), a crash was seen. It was because the idx was -1 and could not be accessed. Now a check is placed to check for idx < 0 and if found change the insertion idx to 0.  Since MD value are always increasing hence the insertion idx should be 0 in this case.

Test Plan: Fix has been tested in F3Demo2015 and is working fine. Row with data is added and rest of the data is preserved.

Reviewers: bert

Reviewed By: bert

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11023