od: Made the VolumeReader VolumeProcessing step release its output after reading
Summary: since the created task was the reader, the output datapack was refcounted until the task was deleted. Changed that so the reader gets out of scope after the execution of the task, but before its deletion.

Test Plan: tested from dip computed volume processing chain. The refcount has decreased as expected.

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11027