od: Fixed two errors in ArrayUdfValReplacer implementation
Summary:
changed call to PosInfo::CubeData::include to PosInfo::CubeData::isValid. The first is true for all traces in the enveloppe, whereas the second is really trace by trace. Also fixed wrong argument for memSet function.
Added a task for writting back undefined values in the output, before writing the datapack on disk.

Test Plan: not all udf samples were removed before that patch, now it is the case.

Reviewers: yuancheng

Reviewed By: yuancheng

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11064