[od] Horizon points on random line sections speedup (reopening OD-1143)
Summary:
(1) Speedup for locked and picked horizon points on random line sections.
New calcDist(.) algorithm is node-based instead of trace-path-based. The
number of nodes of an average random line is orders of magnitude smaller
than the number of traces along its path.
(2) Solution for incorrect update of locked points of a horizon displayed on
sections when dragging a random line. Ayrat showed that points were running
one step behind. This is caused by Ding's modifications to fix crash OD-1050.
The (historical) updating of random line BinIDs path inside the calcDist(.)
function was indeed not a good idea in a parallel context, but Ding's fix
in updatePanelStripPath(.) has not yet reached at the time calcDist(.) is
called, because of getUpdateStageNr()==1 in the preceding return condition.
I could have improved this, but unlike the BinID path, the random line nodes
are already up-to-date in that stage. Since these are used in the new
calcDist() algorithm, Ding's revision r45059 can safely be reverted.

Test Plan:
1) Speed of locked points on section update is now comparable for
the default inline and the default random line. The latter no longer
gets slower according as the length of the random line increases.
Added temporary code to verify whether the intersection points found
are actually present in the trcspath_ typeset of the RandomTrackDisplay.
Only when rotating the random line, an occasional discrepancy occurs.
It always turns out to be the boundary case where fabs(ydiff)==0.5,
which is totally harmless: the new algorithm shows to be equal but faster.
(2) Locked points displayed only at sections on a horizon now remain
intersected all the time with a random line being dragged by the user.
Also checked that the crash fixed by Ding's reverted r45059/OD-1050
did not reoccur.

Reviewers: nanne

Reviewed By: nanne

Subscribers: ding

Differential Revision: https://phabricator.dgbes.com:8443/D11062