od: Reworked progress reporting using ProgressRecorder: Fixes lack of progress update in applying a volume builder step
Summary:
to much information was forwarded from the ProgressRecorder to its class member forwardto_. That let to inconsistent reporting of the task name that was started and finished. The chain executor now gets the ProgressRecorder forwardto_ object to display its status, and it now does it properly. In the meantime all calls to VolProc::ChainOutput::progresskeeper_.setMessage were in vain: The text argument did not get printed in the log file, now it does.
Finally the last VolProc::ChainOutput::storers_ object that remain after the execution of all chunks of the chain get a progressmeter_ so it can nicely report the writing part of the last chunk.

Only con: ProgressMeter::skipProgress is not always usable since it triggers to much interference between tasks. As a result the only alternative way to disable progress in an task that will run sub-tasks is to have this object without name and with an empty uiMessage(). No sure how to do differently, but for sure not all tasks need to get their name printed in user-interface (log or window).

Test Plan: tested by applying a volume builder workflow (dip processing).

Reviewers: kristofer, bert

Reviewed By: bert

Subscribers: nanne, raman, kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D11078