Rework of GLInfo
Summary:
The current implementation of GLInfo did not work on Mac or Windows. The new
implementation does the check just before first render, and if it is terribly
wrong, it will turn off the render.

The update also has the benefit that it queries other things as well (such as shading), which can
be used by the rest of the code.

Test Plan: Tested on mac, but I'm quite sure it will work on other platforms as well.

Reviewers: nanne

Reviewed By: nanne

Subscribers: jaap

Differential Revision: https://phabricator.dgbes.com:8443/D11065