od: Showing inline grids for survey map.
Summary: Showing inline grids for survey maps to give initial sense of survey orientation to the user, using already existing class uiGrid2DMapObject to do this

Test Plan:
But there is a problem, it seems there is a graphics transform problem with the uiLineItem, because setting a width of 200 makes the line visible, other wise the line looks extremely thin and barely visible. This does not happel for uiPolyLineItem

This can even be observed if you do a creeate 2D grid from random line workflow.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11083