od6.0: OD-1119, Active Neural Network(either saved or unsaved) output attributes are not added to 2D Lines.
Summary: The neural network when added to 2D line, was searching for parameters that were set to default value and not specifically for the nla. Apart from this the NN attrib was not being displayed if tried directly while loading the line.

Test Plan: Now the values specific to 2d NLA are being filled and read. Attrib is displayed properly and name is displayed correctly too. 3D is also being loaded correctly.

Reviewers: raman

Reviewed By: raman

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11069