od6.0: Changes to sensitivity of Reload
Summary:
Disabled reload when in tracking mode and when the horizon on disk did not change. For this added a function to EM::IOObjInfo to get the lastmodified timestamp.
Relates to OD-1092

Test Plan: Tracked horizon, reload is greyed out. Added a stored horizon, reload is greyed out. Created a horizon with the unconformity tracker, loaded into the scene, created the same horizon again, now the reload item is active.

Reviewers: raman

Reviewed By: raman

Subscribers: ding

Differential Revision: https://phabricator.dgbes.com:8443/D11104