od6.0: Possible fix to Crash while Zooming out in 2D viewer[OD-587]
Summary:
https://jira.dgbes.com:8443/browse/OD-587?filter=-1

It seems this was actually happening because of a bug inside Qt itslef, which has been fixed in later version of Qt. I have comapared the source code for Qt 5.5.1 and there it seems to be using the correct variable.

So using a compiler directive for Qt version and have put a check on the width of the image

Test Plan: could not reproduce the crash, however the crash itself was rarely reproducible, so I assume this is fixed, atleast it is not going in that qt function

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11137