[od] In-depth plane draggers in z-direction slow in Wheeler scene (OD-1210)
Summary:
In-depth plane draggers in z-direction slow in Wheeler scene (OD-1210). This
is caused by my intelligent draggers of revision r45207/D10679, which used
the survey steps and ranges to scale their dragger movement. This does not
suffice for Z-transformed scenes like Wheeler. Functionality was added to
retrieve those steps and ranges from the scene instead. The bug occurred for
z-slice draggers and the top and bottom planes of the volume dragger. For
the sake of uniformity, the new functionality has also been added to random
line dragging in od, but just not in od6.0 to keep ABI preservation simple.

Test Plan:
In-depth plane dragging of inl/crl/z-slice, random line and volumes works
as before in normal scenes. Dragging of Z-slices and the top and bottom
planes of a volume is no longer slow in transformed scenes.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D11111