od: Made the VolProc::VolumeReader actually read more than the first component
Summary: only the first component was read, because the output datapack of the VolProc::Chain was not aware of the components of the input dataset. Thus made sure all components would be read including their name. Also added a possibility to read only some of the components

Test Plan: tested from FaultLikelihood processing (plugins)

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D11140