od: made the "update" target of od also update the plugins if applicable
Summary:
one may specify the path to the closed source plugins with "PLUGIN_DIR", for instance to have all targets into one single Visual Studio project. I increased the scope of the update target so it will:
1- Update the sources of the closed source plugins
2- Update their externals too.

Test Plan: tested with a valid and not-emtpy PLUGIN_DIR on windows, and with an empty PLUGIN_DIR on linux. Got what I expected.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D11144