[od6.0] Added environment variable to force crashdump. Added machine-readable crashdump after the human readable.
Summary: I wish to make the crashdumps machine readable, and hence I want the output with the -m output as well as the normal human readable.

Test Plan:
Tested on Linux, and works fine. NOT tested on Windows. The stackwalker application has been checked that it uses the -m flag, but dumps cannot easily be done on Windows. That is also why the flag is added, so it can be easily provoked by setting an environment variable.

The old CRASH flag crashes too early, before the crashdump facility is installed.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11156