[OD]OD-670 Import 3D horizon: horizon is only partially imported when surface grid is different from OpendTect survey grid
Summary:
After investigation, the problem was caused by function includes( BinID ) in TrcKeySampling in which it does always considering steps. 
But in this case, we don't need filtered out data by steps. We do only need know if the sample is in the survey range.
Solution:
Added extra bool variable in includes function.

Test Plan: Tested in the survey:/dsk/d35/surveys/OD670 by the steps mentioned in the Jira issue. It works.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D11096