od6.0: Batch program start windows should not close on OK
Summary:
Made BatchProcDlgs return false, ie don't close them on OK. This makes it possible to easily start a batchprogram twice with a small change in parameters,
Added a uiString for the general start error message.
Other batch windows, which are not a subclass have to change as well.

Test Plan: Tested a few windows, and all work as expected.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11183