od: Added RegularSeisDataPack::clone tool
Summary: as discussed with Kristofer, to avoid editing the input datapack of a volume processing step, one should copy it and edit the copied datapack. So added a clone tool to perform step 1.

Test Plan: tested reading of class members of the copied datapack after the source was deleted. All good.

Reviewers: bert

Reviewed By: bert

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D11189