od6.0: Better MMP window caption and option to save nrinlperjob
Summary:
Every IOPar should have a Target. This string is then used to set the window caption. No window title is needed.
Also added a save as default for nrofinlperjob. Not sure if the moment of save is OK.
Work in progress for OD-339 and OD-1232

Test Plan:
Looked at a few windows for attribute and volumebuilder processes. Title looks OK.
nrofinlperjob is saved in settings file when box is checked.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11199