od6.0: Fixed zstep problem
Summary:
See OD-1012. The field was simply missing and it does not harm if it's there when no extra Z is selected.
Regarding the extraz field, the label was wrong. Not sure why, I guess a problem in the usage of mFromUiStringTodo. As uiSelZRange does not accept uiStrings yet, the label can be given as a const char*.

Test Plan: gui looks ok now. Did not test the execution, but nothing has changed there.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11200