[OD6.0]OD-670 Import 3D horizon: horizon is only partially imported when surface grid is different from OpendTect survey grid
Summary:
This revision is ported from the revision D11096 which was accepted
After investigation, the problem was caused by function includes( BinID ) in TrcKeySampling in which it does always considering steps.
But in this case, we don't need filtered out data by steps. We do only need know if the sample is in the survey range.
Solution:
In order to keep ABI, added extra functions in trckeysampling.

Test Plan: Tested in the survey:/dsk/d35/surveys/OD670 by the steps mentioned in the Jira issue. It works.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D11169