od: Almost fixes  empty markers list from unconformity tracker and well manager interface
Summary: many static pointers were not copied into local BufferString, leading to names corruption. Also there was a mistake between a level ID and its index in the set. That created a complete renumbering of the IDs, so the object in memory did not match the database.

Test Plan: retrieval of marker depths in unconformity tracker is fine now. "Edit Well Markers" dialog works better but not yet 100%. First time is OK, next times the name still get corrupted.

Reviewers: helene

Reviewed By: helene

Subscribers: raman, bert

Differential Revision: https://phabricator.dgbes.com:8443/D11222