od: Fixed corruption of level name for well markers
Summary: uiStratLevelSel::getLevelName() was returning a pointer to a local object, which was thus corrupted as it got deleted when going out of scope. This local object was previously a pointer to a class member. Fixed it by returning a new BufferString. Please confirm this is the way to go, or if I shall make use of a static object.

Test Plan: marker names are no longer corrupted when re-opening the Edit Well Markers dialog several times.

Reviewers: raman, bert

Reviewed By: bert

Subscribers: bert, kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D11223