[OD]OD-630,OD-631 manual tracking node issue.
Summary: This is temporal revision for rough review to check if the issue is handled in the correct way. It will be abandoned after fully testing.

Test Plan:
For the time being I can not do the fully testing due to current od truck version.
I have donw part of testing, it works.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D11202