od6.0: Fixed a crash during reading of seismic in VolumeProcessing
Summary:
The ArrayFiller class was using a reference to a HiddenParam object. The crash may happen if the SequentialReader was deleted before the reading was finished, then the reference would point to a removed object. May explain the crash seen by Paul in Fault Fracture basement processing.
Replaced the reference by a copyied object.

Test Plan: no more crash at the end of the reading step while debugging the crash reported by Paul.

Reviewers: jaap

Reviewed By: jaap

Subscribers: bert

Differential Revision: https://phabricator.dgbes.com:8443/D11256