od: OD-963, Application of multiple markers for flattening of layer model
Summary: Layer model was being flattened right now on the basis of single marker. Rest of the markers were not displayed. Now the scene will be flattened on the basis of the marker of your choice and rest of the selected markers will also be displayed

Test Plan: Tested in F3_Demo_2012 and is working fine

Reviewers: satyaki

Reviewed By: satyaki

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11262