od6.0: Missing ft to meter conversion
Summary: Another bug in the volume estimation. This time for surveys with depth in feet. The function should return a m3, so the depth in feet should be converted to meter.

Test Plan: Calculations look OK now. Still a difference between value I calculate myself and OD. Probably a boundary issue.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D11229