[OD]OD-1383 Parts of horizon auto-tracked on section become invisible if the section is moved
Summary:
After done investigation, the problem is caused by that the notification: movementnotifier in geoelement is not invoked when doing single click tracking and manual draw tracking. 
Solution: added necessary code to set proper blockCallBacks in order to trigger the changes of horizon geometry.
Note: for 2d horizon it is not needed to invoke this notifier.

Test Plan: Tested in F3-demo as described in the Jira issue OD-1383. It works fine.

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D11227