od6.0: Give unique jobname to copy job
Summary: Every jobname was the same. So when starting a second copy job, it overwrites the first one. Not sure what happens on Windows, but I think it's the cause of the problem described in OD-1460

Test Plan: Started 2 copy jobs. Par file name is unique and both jobs finish correctly (on Linux). Windows needs to be tested.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11284