od6.0: Fixed copy of certain component with nulls added
Summary:
If you copy only the cross-line component of a steeringcube and add nulls, all samples will be null. See OD-1459
Also fixed the Type in the omf file in this case. It was still Steering, whereas a single component can never be zero.

To explain my fix in SeisCubeCopier::doProc:
I'm not sure when veltype_ is set in the SeisCubeCopier, so I left that implementation untouched. If veltype==0, and compnr_!=-1, it looks like the trc always has 1 component.

Test Plan: Copied crl dip component and added nulls. Result is now as expected.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11285