od6.0: OD-1404 Crash while changing survey while SynthRock window open
Summary: The crash occured when we tried to change the survey when the dialog for opening or saving model data was launched but not closed. The window was being deleted twice.

Test Plan: The popped up dialog is now made non modal by changing the parent. This is a temporary fix for 6.0. Actual fix has been done in OD-master.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11316