od6.0: OD-1366, Crash in setSectionDisplayRestoreForAllHors
Summary: There was no description of crash, I did the following checks. I added three horizons in the scene, switched the view mode between At Sections and In Full. Saved the display and restored it after closing the OD. But still everything is working fine. I have placed a check for null pointer in the function.

Test Plan: The display is still fine

Reviewers: ranojay

Reviewed By: ranojay

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11317