od6.0: Corrected the chunking of the volume processing
Summary: the entire chain size was wrongly doubled. First there is no need to take into account the output datapack of a previous is only one chunk will be used. Secondly the extra size amount to the output datapack size, not the memory usage of the entire chain. Here I multiply this output size arbitrarily by 3, the total maximum of output components were are using. There are no tools in od6.0 that can tell me how many components the output of the chain will have.

Test Plan: tested on Nexen data. The number of chunks is reduced from 19 to 12. The process uses up to a large part of the free memory but not all.

Reviewers: bert

Reviewed By: bert

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D11319